feat(batch-4): partial landings (roadmap+OpenAPI yaml, snapshot replay-protection fields, mem max_linear_memory config, wasi-gpu/tenant tests, fuzz scaffolds, OpenAI shim scaffold); orchestrator fix-ups for compile

+++ crates/tensor-wasm-exec/tests/jit_dispatch_per_store.rs
@@ -48,7 +48,7 @@ fn make_engine() -> Engine {
 }
 
 fn instance_state(tenant: u64, id: u64) -> InstanceState {
-    InstanceState::new(TenantId(tenant), InstanceId(id))
+    InstanceState::new(TenantId(tenant), InstanceId(id as u128))
 }
 
 /// Two stores from the SAME linker must not share an arena.
